Integrando servico REST

+++ controllers/pagamentos.js
@@ -43,18 +43,17 @@ module.exports = function(app){
 
 
     app.post('/pagamentos/pagamento',function(req, res){
-
         req.assert("pagamento.forma_de_pagamento",
         "Forma de pagamento obrigatorio").notEmpty();
         req.assert("pagamento.valor",
         "Valor obrigatorio e deve ser decimal").notEmpty().isFloat();
 
-        var erros = req.validationErrors();
-        if(erros){
-            console.log("erro de validadcao");
-            res.status(400).send(erros);
-            return;
-        }
+            var erros = req.validationErrors();
+            if(erros){
+                console.log("erro de validadcao");
+                res.status(400).send(erros);
+                return;
+            }
 
         var pagamento = req.body["pagamento"];
         pagamento.status = 'Criado';
@@ -66,44 +65,43 @@ module.exports = function(app){
             if(err){
                 console.log("erro ai inserir");
                 res.status(500).send(err);
-            }else{
-            pagamento.id = result.insertId;
+            }
+            else
+            {
+                pagamento.id = result.insertId;
 
                 if(pagamento.forma_de_pagamento == 'cartao'){
                     var cartao = req.body['cartao'];
-
                     var clienteCartoes = new app.servicos.clienteCartoes();
+
                         clienteCartoes.autoriza(cartao, function(erro, request, response, retorno){
-                            console.log(retorno);
-                             res.status(201).json(retorno);
+                            if(erro){
+                                res.status(400).send(erro);
+                                return;
+                            }
+                             console.log(retorno);
+                             res.location('/pagamentos/pagamento/' + pagamento.id);
+                                var response = {
+                                    dados_do_pagamento: pagamento,
+                                    cartao: retorno,
+                                    links: [
+                                        {
+                                            href:"http://localhost:3000/pagamentos/pagamento/" + pagamento.id,
+                                            rel:"CONFIRMAR",    
+                                            method:"PUT"
+                                        },
+                                        {
+                                            href:"http://localhost:3000/pagamentos/pagamento/" + pagamento.id,
+                                            rel:"CANCELAR",    
+                                            method:"DELETE"
+                                        }
+                                    ]
+                                }
+                            res.status(201).json(retorno);
                             return;
                         });
-                }else{
-                    res.location('/pagamentos/pagamento/' + pagamento.id);
-            
-                    var response = {
-                        dados_do_pagamento: pagamento,
-                        links: [
-                            {
-                                href:"http://localhost:3000/pagamentos/pagamento/" + pagamento.id,
-                                rel:"CONFIRMAR",    
-                                method:"PUT"
-                            },
-                            {
-                                href:"http://localhost:3000/pagamentos/pagamento/" + pagamento.id,
-                                rel:"CANCELAR",    
-                                method:"DELETE"
-                            }
-                        ]
-                    }
-
-                res.status(201).json(response);
                 }
             }
-
-            
-
-           
         });
     });
 }
